+++ moban/plugins/strategy.py
